refactor(OnImportAndRemove): 🔄 restructure assembly loading for PowerShell 7+
* Removed the `LoadContext` class from the `LoadContext.cs` file.
* Integrated assembly loading directly into `OnModuleImportAndRemove` for better clarity and maintainability.
* Enhanced support for both PowerShell 5.1 and 7+ with clear separation of implementation details.
